refactored server and sockets relates #55
